use friend_test for test
