+++ src/server/pegasus_server_write.h
@@ -58,6 +58,7 @@ private:
 private:
     friend class pegasus_server_write_test;
     friend class pegasus_write_service_test;
+    friend class pegasus_write_service_impl_test;
 
     std::unique_ptr<pegasus_write_service> _write_svc;
     std::vector<put_rpc> _put_rpc_batch;
