+++ src/server/pegasus_write_service.h
@@ -157,6 +157,7 @@ private:
 
 private:
     friend class pegasus_write_service_test;
+    friend class pegasus_write_service_impl_test;
     friend class pegasus_server_write_test;
 
     pegasus_server_impl *_server;
